feat: Convert elements of path variable to edge references
A path column is an array of `graphid`. It should actually be an array
of edges, but the size of edge is quite large. So, using the array
causes performance degradation. Therefore, `edgeref` type is introduced.

`edgeref` consists of a `relid` and a `tid`. You can get information of
an edge through `edgeref` which points the edge whenever you need.

`relid` is an index of the relation array that is in `EState`. And `tid`
is the *self item pointer* of an edge heap tuple.

A `edgeref()` function is added to the `targetlist` of VLE in transform
phase. And it is used as a placeholder. After inheritance of the edge
was processed in prepunion, the first argument is replaced to a real
`relid` while creating the VLE join plan.

Following functions are introduced.

* `edgeref()`: converts (`relid`, `tid`) pair to `edgeref`
* `edgeref_in/out()`: input/output function for `cstring`
* `edgerefprop()`: fetch the edge heap tuple and return the property
* `edgerefrow()`: fetch the edge heap tuple and return the row for it

(cherry picked from commit 8450b9e6a088c42f34cf697ad22e28ec2e0dae83)