fix: addendum to LIMIT clause causes VLEs to crash (AG-254)
This is an additional fix for the issue AG-254, commit id -

`commit 4866bfe296cd40fec54aac64b9afdb22c9f49e6b`

Upon further testing it was found that if there are multiple VLABELs or
ELABELs, VLEs will crash the session if the LIMIT clause is used. This
happens whether those labels are used or not. This fix is to address
this issue.

The issue is caused by the PlanState node's chgParam values being set,
but due to the LIMIT clause, they are not processed. The solution is to
free them and then NULL them out.

The function `freePlanStateChgParam` is added as part of the cleanup
routine to free the chgParam values in the PlanState node.

The regression sql script and the expected output are adjusted for this
addition.

Note: This fix compiles and passes all install checks

-jrg

(cherry picked from commit c8fadc0310f0d63e603a365e443f0db45cd33933)