fix: ExecSetGraph fails on SET commands
This fix addresses AG-279 -

ERROR: type with OID 2139062143 does not exist

The problem occurs in SET commands when there are more than 2 values
in a row to set - two or less will always succeed.

The fix is to `findAndReflectNewestValue` to always make a defensive
copy for a retrieved hash table value prior to it being stored in the
slot. Then freeing any value in the slot prior to setting it.

(cherry picked from commit 264495e5c2d1d3d4d4c2f10a23721533a5814e35)